requirements.txt
shortuuid
django-phone-field
phonenumbers
pyotp